+++ lib/src/bindings/events.dart
@@ -1,3 +1,6 @@
+// Copyright (c) 2017, Anatoly Pulyaevskiy. All rights reserved. Use of this source code
+// is governed by a BSD-style license that can be found in the LICENSE file.
+
 @JS()
 library node_interop.bindings.events;
 
@@ -7,4 +10,5 @@ import 'package:js/js.dart';
 @JS()
 abstract class EventEmitter {
   external void on(String eventName, listener);
+  external void removeAllListeners([String eventName]);
 }