+++ lib/src/http_server.dart
@@ -1,20 +1,34 @@
 // Copyright (c) 2017, Anatoly Pulyaevskiy. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 import 'dart:async';
-import 'dart:convert';
 import 'dart:io' as io;
 
 import 'package:js/js.dart';
+import 'package:js/js_util.dart';
 
 import 'bindings/globals.dart';
 import 'bindings/http.dart' as nodeHTTP;
+import 'http_headers.dart';
 import 'internet_address.dart';
-import 'package:js/js_util.dart';
-import 'util.dart';
+import 'streams.dart';
 
-export 'dart:io' show HttpStatus, HttpHeaders, ContentType;
+export 'dart:io'
+    show HttpStatus, HttpHeaders, ContentType, Cookie, HttpException;
 
 final nodeHTTP.HTTP _http = require('http');
+
+class _HttpConnectionInfo implements io.HttpConnectionInfo {
+  @override
+  final int localPort;
+
+  @override
+  final InternetAddress remoteAddress;
+
+  @override
+  final int remotePort;
+
+  _HttpConnectionInfo(this.localPort, this.remoteAddress, this.remotePort);
+}
 
 class HttpServer extends Stream<io.HttpRequest> implements io.HttpServer {
   @override
@@ -60,6 +74,7 @@ class HttpServer extends Stream<io.HttpRequest> implements io.HttpServer {
       response.statusCode = io.HttpStatus.SERVICE_UNAVAILABLE;
       response.end();
     }
+    _controller.add(new HttpRequest(request, response));
   }
 
   Future<HttpServer> _bind() {
@@ -115,289 +130,185 @@ class HttpServer extends Stream<io.HttpRequest> implements io.HttpServer {
   }
 }
 
-class _HttpConnectionInfo implements io.HttpConnectionInfo {
-  @override
-  final int localPort;
-
-  @override
-  final InternetAddress remoteAddress;
-
-  @override
-  final int remotePort;
-
-  _HttpConnectionInfo(this.localPort, this.remoteAddress, this.remotePort);
-}
-
-/// HTTP headers which can not be modified.
-///
-/// All mutation methods of this class throw [UnsupportedError].
-class ImmutableHttpHeaders implements io.HttpHeaders {
-  final int _port;
-  final dynamic _headers;
-
-  ImmutableHttpHeaders(this._headers, this._port);
-
-  @override
-  bool get chunkedTransferEncoding =>
-      getProperty(_headers, io.HttpHeaders.TRANSFER_ENCODING) == 'chunked';
-
-  @override
-  void set chunkedTransferEncoding(bool chunked) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  int get contentLength {
-    var value = getProperty(_headers, io.HttpHeaders.CONTENT_LENGTH);
-    if (value != null) return int.parse(value);
-    return 0;
-  }
-
-  @override
-  set contentLength(int length) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  io.ContentType get contentType {
-    if (_contentType != null) return _contentType;
-    String value = getProperty(_headers, io.HttpHeaders.CONTENT_TYPE);
-    if (value == null || value.isEmpty) return null;
-    var types = value.split(',');
-    _contentType = io.ContentType.parse(types.first);
-    return _contentType;
-  }
-
-  io.ContentType _contentType;
-
-  @override
-  set contentType(io.ContentType type) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  DateTime get date {
-    String value = getProperty(_headers, io.HttpHeaders.DATE);
-    if (value == null || value.isEmpty) return null;
-    try {
-      return io.HttpDate.parse(value);
-    } on Exception catch (e) {
-      return null;
-    }
-  }
-
-  @override
-  set date(DateTime date) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  DateTime get expires {
-    String value = getProperty(_headers, io.HttpHeaders.EXPIRES);
-    if (value == null || value.isEmpty) return null;
-    try {
-      return io.HttpDate.parse(value);
-    } on Exception catch (e) {
-      return null;
-    }
-  }
-
-  @override
-  set expires(DateTime expires) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  String get host => getProperty(_headers, io.HttpHeaders.HOST);
-
-  @override
-  set host(String host) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  DateTime get ifModifiedSince {
-    String value = getProperty(_headers, io.HttpHeaders.IF_MODIFIED_SINCE);
-    if (value == null || value.isEmpty) return null;
-    try {
-      return io.HttpDate.parse(value);
-    } on Exception catch (e) {
-      return null;
-    }
-  }
-
-  @override
-  set ifModifiedSince(DateTime ifModifiedSince) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  bool get persistentConnection {
-    var connection = getProperty(_headers, io.HttpHeaders.CONNECTION);
-    return (connection == 'keep-alive');
-  }
-
-  @override
-  set persistentConnection(bool persistentConnection) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  List<String> operator [](String name) {
-    throw new UnimplementedError();
-  }
-
-  @override
-  void add(String name, Object value) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  void clear() {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  void forEach(void f(String name, List<String> values)) {
-    throw new UnimplementedError();
-  }
-
-  @override
-  void noFolding(String name) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  void remove(String name, Object value) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  void removeAll(String name) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  void set(String name, Object value) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  String value(String name) {
-    throw new UnimplementedError();
-  }
-
-  @override
-  set port(int port) {
-    throw new UnsupportedError('HttpHeaders are immutable');
-  }
-
-  @override
-  int get port => _port;
-}
-
-class HttpRequest extends Stream<List<int>> implements io.HttpRequest {
-  final nodeHTTP.IncomingMessage _nativeRequest;
+/// Server side HTTP request object which delegates IO operations to
+/// Node's native representations.
+class HttpRequest extends ReadableStream<List<int>> implements io.HttpRequest {
   final nodeHTTP.ServerResponse _nativeResponse;
 
-  HttpRequest._(this._nativeRequest, this._nativeResponse);
+  HttpRequest(nodeHTTP.IncomingMessage nativeRequest, this._nativeResponse)
+      : super(nativeRequest, convert: (chunk) => new List.unmodifiable(chunk));
+
+  nodeHTTP.IncomingMessage get nativeStream => super.nativeStream;
 
   @override
   io.X509Certificate get certificate => throw new UnimplementedError();
 
   @override
   io.HttpConnectionInfo get connectionInfo {
-    var netAddress = _nativeRequest.socket.address();
-    var address = new InternetAddress(netAddress.address, null);
-    return new _HttpConnectionInfo(netAddress.port, address, null);
+    var socket = nativeStream.socket;
+    var address = new InternetAddress(socket.remoteAddress, null);
+    return new _HttpConnectionInfo(
+        socket.localPort, address, socket.remotePort);
   }
 
   @override
   int get contentLength => headers.contentLength;
 
   @override
-  List<io.Cookie> get cookies => throw new UnimplementedError();
-
-  @override
-  io.HttpHeaders get headers => new ImmutableHttpHeaders(
-      _nativeRequest.headers, connectionInfo.remotePort);
-
-  StreamController<List<int>> _controller;
-  @override
-  StreamSubscription<List<int>> listen(void onData(List<int> event),
-      {Function onError, void onDone(), bool cancelOnError}) {
-    assert(_controller == null);
-    _controller = new StreamController();
-
-    _nativeRequest.on('close', allowInterop(() {
-      _controller.close();
-    }));
-    _nativeRequest.on('data', allowInterop((Iterable<int> chunk) {
-      _controller.add(new List.unmodifiable(chunk));
-    }));
-    return _controller.stream.listen(onData,
-        onError: onError, onDone: onDone, cancelOnError: cancelOnError);
+  List<io.Cookie> get cookies {
+    if (_cookies != null) return _cookies;
+    _cookies = new List<io.Cookie>();
+    List<String> values = headers[io.HttpHeaders.SET_COOKIE];
+    if (values != null) {
+      values.forEach((value) {
+        _cookies.add(new io.Cookie.fromSetCookieValue(value));
+      });
+    }
+    return _cookies;
   }
 
+  List<io.Cookie> _cookies;
+
   @override
-  String get method => _nativeRequest.method;
+  io.HttpHeaders get headers =>
+      _headers ??= new RequestHttpHeaders(nativeStream);
+  io.HttpHeaders _headers;
+
+  @override
+  String get method => nativeStream.method;
 
   @override
   bool get persistentConnection => headers.persistentConnection;
 
   @override
-  String get protocolVersion => _nativeRequest.httpVersion;
+  String get protocolVersion => nativeStream.httpVersion;
 
   @override
-  Uri get requestedUri => throw new UnimplementedError();
+  Uri get requestedUri {
+    if (_requestedUri == null) {
+      var socket = nativeStream.socket;
+
+      var proto = headers['x-forwarded-proto'];
+      var scheme;
+      if (proto != null) {
+        scheme = proto.first;
+      } else {
+        var isSecure = getProperty(socket, 'encrypted') ?? false;
+        scheme = isSecure ? "https" : "http";
+      }
+      var hostList = headers['x-forwarded-host'];
+      String host;
+      if (hostList != null) {
+        host = hostList.first;
+      } else {
+        hostList = headers['host'];
+        if (hostList != null) {
+          host = hostList.first;
+        } else {
+          host = "${socket.localAddress}:${socket.localPort}";
+        }
+      }
+      _requestedUri = Uri.parse("$scheme://$host$uri");
+    }
+    return _requestedUri;
+  }
+
+  Uri _requestedUri;
 
   @override
   io.HttpResponse get response =>
-      _response ??= new HttpResponse._(_nativeResponse);
-  io.HttpResponse _response;
+      _response ??= new HttpResponse(_nativeResponse);
+  io.HttpResponse _response; // ignore: close_sinks
 
   @override
   io.HttpSession get session => throw new UnsupportedError(
       'Sessions are not supported by Node HTTP server.');
 
   @override
-  Uri get uri => Uri.parse(_nativeRequest.url);
+  Uri get uri => Uri.parse(nativeStream.url);
 }
 
-class HttpResponse extends io.HttpResponse {
-  final nodeHTTP.ServerResponse _nativeResponse;
+class HttpResponse extends NodeIOSink implements io.HttpResponse {
+  HttpResponse(nodeHTTP.ServerResponse nativeResponse) : super(nativeResponse);
+
+  nodeHTTP.ServerResponse get nativeStream => super.nativeStream;
 
   @override
-  Encoding encoding;
-
-  HttpResponse._(this._nativeResponse);
+  bool get bufferOutput => throw new UnimplementedError();
 
   @override
-  void add(List<int> data) {
+  set bufferOutput(bool buffer) {
     throw new UnimplementedError();
   }
 
   @override
-  void addError(Object error, [StackTrace stackTrace]) {
+  int get contentLength => throw new UnimplementedError();
+
+  @override
+  set contentLength(int length) {
     throw new UnimplementedError();
   }
 
   @override
-  Future addStream(Stream<List<int>> stream) {
+  Duration get deadline => throw new UnimplementedError();
+
+  @override
+  set deadline(Duration value) {
     throw new UnimplementedError();
+  }
+
+  @override
+  bool get persistentConnection => headers.persistentConnection;
+
+  @override
+  set persistentConnection(bool persistent) {
+    headers.persistentConnection = persistent;
+  }
+
+  @override
+  String get reasonPhrase => nativeStream.statusMessage;
+  set reasonPhrase(String phrase) {
+    if (nativeStream.headersSent) throw new StateError('Headers already sent.');
+    nativeStream.statusMessage = phrase;
+  }
+
+  @override
+  int get statusCode => nativeStream.statusCode;
+
+  @override
+  set statusCode(int code) {
+    if (nativeStream.headersSent) throw new StateError('Headers already sent.');
+    nativeStream.statusCode = code;
   }
 
   @override
   Future close() {
-    _nativeResponse.end();
-    return new Future.value();
+    _headers.finalize();
+    return super.close();
   }
 
   @override
-  io.HttpConnectionInfo get connectionInfo => throw new UnimplementedError();
+  io.HttpConnectionInfo get connectionInfo {
+    var socket = nativeStream.socket;
+    var address = new InternetAddress(socket.remoteAddress, null);
+    return new _HttpConnectionInfo(
+        socket.localPort, address, socket.remotePort);
+  }
 
   @override
-  List<io.Cookie> get cookies => throw new UnimplementedError();
+  List<io.Cookie> get cookies {
+    if (_cookies != null) return _cookies;
+    _cookies = new List<io.Cookie>();
+    List<String> values = headers[io.HttpHeaders.SET_COOKIE];
+    if (values != null) {
+      values.forEach((value) {
+        _cookies.add(new io.Cookie.fromSetCookieValue(value));
+      });
+    }
+    return _cookies;
+  }
+
+  List<io.Cookie> _cookies;
 
   @override
   Future<io.Socket> detachSocket({bool writeHeaders: true}) {
@@ -405,38 +316,14 @@ class HttpResponse extends io.HttpResponse {
   }
 
   @override
-  Future get done => throw new UnimplementedError();
-
-  @override
-  Future flush() {
-    throw new UnimplementedError();
-  }
-
-  @override
-  io.HttpHeaders get headers => throw new UnimplementedError();
+  io.HttpHeaders get headers =>
+      _headers ??= new ResponseHttpHeaders(nativeStream);
+  ResponseHttpHeaders _headers;
 
   @override
   Future redirect(Uri location, {int status: io.HttpStatus.MOVED_TEMPORARILY}) {
-    throw new UnimplementedError();
-  }
-
-  @override
-  void write(Object obj) {
-    throw new UnimplementedError();
-  }
-
-  @override
-  void writeAll(Iterable objects, [String separator = ""]) {
-    throw new UnimplementedError();
-  }
-
-  @override
-  void writeCharCode(int charCode) {
-    throw new UnimplementedError();
-  }
-
-  @override
-  void writeln([Object obj = ""]) {
-    // TODO: implement writeln
+    statusCode = status;
+    headers.set("location", location);
+    return close();
   }
 }