+++ test/test_util_test.dart
@@ -25,4 +25,8 @@ void main() {
 }
 
 @JS('Error')
-abstract class JsError {}
+abstract class JsError {
+  external String get code;
+  external String get message;
+  external String get stack;
+}